// frontend/src/generated/api/models/PaginationMeta.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PaginationMeta = {
    /**
     * 条件に一致する総アイテム数
     */
    totalItems: number;
    /**
     * 総ページ数
     */
    totalPages: number;
    /**
     * 現在のページ番号 (1始まり)
     */
    currentPage: number;
    /**
     * 1ページあたりのアイテム数
     */
    limit: number;
};