// frontend/src/generated/api/models/User.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type User = {
    /**
     * ユーザーの一意なID
     */
    readonly id: string;
    /**
     * ユーザー名 (表示名)
     */
    name: string;
    /**
     * アカウント作成日時
     */
    readonly createdAt: string;
    /**
     * アカウント更新日時
     */
    readonly updatedAt: string;
};